Renamed files and constants to denote data/resource
