+++ internal/provider/data_usernames.go
@@ -23,17 +23,17 @@ import (
 )
 
 const (
-	roleNamesRoleNamesKey = "role_names"
-	roleNamesIDValue      = "splunkconfig_role_names"
+	userNamesUserNamesKey = "user_names"
+	userNamesIDValue      = "splunkconfig_user_names"
 )
 
-func resourceRoleNames() *schema.Resource {
+func dataUserNames() *schema.Resource {
 	return &schema.Resource{
-		Description: "Return Role Names from the Splunk Configuration",
-		ReadContext: resourceRoleNamesRead,
+		Description: "Return User Names from the Splunk Configuration",
+		ReadContext: resourceUserNamesRead,
 		Schema: map[string]*schema.Schema{
-			roleNamesRoleNamesKey: {
-				Description: "List of Role Names in the Splunk Configuration",
+			userNamesUserNamesKey: {
+				Description: "List of User Names in the Splunk Configuration",
 				Type:        schema.TypeList,
 				Computed:    true,
 				Elem:        &schema.Schema{Type: schema.TypeString},
@@ -42,11 +42,11 @@ func resourceRoleNames() *schema.Resource {
 	}
 }
 
-func resourceRoleNamesRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+func resourceUserNamesRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	suite := meta.(config.Suite)
 
-	d.SetId(roleNamesIDValue)
-	if err := d.Set(roleNamesRoleNamesKey, suite.ExtrapolatedRoles().RoleNames()); err != nil {
+	d.SetId(userNamesIDValue)
+	if err := d.Set(userNamesUserNamesKey, suite.Users.Names()); err != nil {
 		return diag.FromErr(err)
 	}
